ASK-826 Не удаляются демо данные вручную в интерфейсе

+++ src/AdminCabinet/Controllers/AsteriskManagersController.php
@@ -154,6 +154,10 @@ class AsteriskManagersController extends BaseController
      */
     public function deleteAction(string $amiId = ''): void
     {
+        if ($amiId === '') {
+            return;
+        }
+
         $manager = AsteriskManagerUsers::findFirstByid($amiId);
         if ($manager !== null) {
             $manager->delete();
